+++ actionpack/lib/action_controller/railtie.rb
@@ -21,10 +21,11 @@ module ActionController
     initializer "action_controller.parameters_config" do |app|
       options = app.config.action_controller
 
-      ActionController::Parameters.permit_all_parameters = options.delete(:permit_all_parameters) { false }
       ActionController::Parameters.action_on_unpermitted_parameters = options.delete(:action_on_unpermitted_parameters) do
         (Rails.env.test? || Rails.env.development?) ? :log : false
       end
+      ActionController::Parameters.always_permitted_parameters = options.delete(:always_permitted_parameters) { %w( controller action ) }
+      ActionController::Parameters.permit_all_parameters = options.delete(:permit_all_parameters) { false }
     end
 
     initializer "action_controller.set_configs" do |app|